requirements.txt
fastapi
uvicorn
python-multipart
jinja2
python-dotenv
PyMuPDF
pdfplumber
simple-ats
requests
google-generativeai
gunicorn 
markdown
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]